Formatting and config_flow.py changes
