+++ custom_components/mqtt_vacuum_camera/utils/vacuum/mqtt_vacuum_services.py
@@ -1,5 +1,5 @@
 """Collection of services for the vacuums and camera components.
-Version: 2025.3.0b0
+Version: 2025.3.0b1
 Autor: @sca075"""
 
 from functools import partial
